admin, getadmin: move deluser to admin
also a hoOg cleanup again
